backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
brotli==1.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cssselect2==0.8.0
docstring_parser==0.17.0
email-validator==2.1.2
fastapi==0.121.1
fonttools==4.60.1
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-aiplatform==1.127.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-genai==1.50.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Markdown==3.10
numpy==2.3.4
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pluggy==1.6.0
PyJWT==2.10.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydyf==0.11.0
Pygments==2.19.2
pyphen==0.17.2
pytest==9.0.0
pytest-asyncio==1.3.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
zopfli==0.4.0
